2단계 - 코드 리뷰 반영
